Update pyoneer.py
Added a catch inside asyncio for loop.  During execution against files certain file 'objects', IE archives without extensions or links to folders, that aren't readable as flat files would cause an exception in asyncio and cause all workers to exit. The catch will allow the single worker to exit and keep the other running and add the worker back to the pool.
Added Files/min stats to on-screen output.